requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-CORS==4.0.0
Werkzeug==3.1.4
gunicorn==21.2.0
psycopg2-binary>=2.9.0
cloudinary>=1.36.0
python-dotenv>=1.0.0
twilio>=8.0.0

